Retain custom average period through three-inside serialization
ThreeInsideUp/DownIndicator stored the average period only inside the transient harami delegate, so a reflection round trip rebuilt the indicator with the default period. Keep the period in a non-transient field, document the default in the class Javadoc, cover custom-period serialization fixtures, and move the large-window head-advance regression into the owning VolumeIndicatorTest.
